Move functions from reducers to model
